+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added WDIO tests.
+
 ### Changed
 * Updated condition for `Secondary Nav item` to scroll when not visible.